requirements.txt
datasets 
typer
huggingface_hub 
kaggle 
pandas 
pyarrow
pymongo
dotenv
google.genai
pathlib
pyyaml
markdown 
pdfkit
python-docx
docx2pdf